CAX8R 单元质量阵直接转化为集中质量阵的对角形式

+++ src/cpp/Solver.cpp
@@ -399,7 +399,7 @@ void CG_alpha::G_alpha_Intregration(CLoadCaseData& Load, int i_load)
 		His_Output->OutputHisMessage(t, dis, vel, acc, N_His_Freedom, num_freedom);
 
 		// Tecplot Output
-		if (fmod(Tec_Count, Ani_Interval) == 0) {
+		if (fmod((double)Tec_Count, (double)Ani_Interval) == 0) {
 			cout << "Output Tecplot and Paraview, Time =  " << t << endl;
 			Tecplot_Output->OutputTecplot(t, dis);
 			Paraview_Output->OutputVTK(t, dis);
